requirements.txt
Flask==3.0.3
Flask-Session==0.5.0
Werkzeug==3.0.1
gunicorn==23.0.0
Flask-Mail==0.9.1
itsdangerous==2.1.2
yagmail==0.15.293
python-dotenv==1.0.0
psycopg2-binary==2.9.9